Fixed music not playing and navBar compatibility
Music was not playing mainly due to the navbar buttons not being compatible with all of the possible audio states that the audio DOM element can have while playing the music. This has been fixed, I would like to have a grayed out button for the pre-music phase (on refresh) but that is not working.